Reworked hn-clone to use react-router:
hn-clone is now fully using react-router
