+++ index.html
@@ -15,7 +15,7 @@
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap-theme.min.css" integrity="sha384-fLW2N01lMqjakBkx3l/M9EahuwpSfeNvV63J5ezn3uZzapT0u7EYsXMjQV+0En5r" crossorigin="anonymous">
 
     <!-- Custom styles for this template -->
-    <link href="public/css/custom.css" rel="stylesheet">
+    <link href="/public/css/custom.css" rel="stylesheet">
 
 
     <!-- jQuery (necessary for Bootstrap's JavaScript plugins) -->
@@ -45,14 +45,14 @@
           <span class="icon-bar"></span>
           <span class="icon-bar"></span>
           </button>
-          <a class="navbar-brand" href="#">Hacker News Clone</a>
+          <a class="navbar-brand" href="/">Hacker News Clone</a>
         </div>
       </div>
     </nav>
 
     <div id="app"></div>
 
-    <script src="bundle.js"></script>
+    <script src="/bundle.js"></script>
 
   </body>
 </html>
